requirements.txt
aerich==0.7.1
aiodns==3.0.0
aiofiles==22.1.0
aiogram==2.23.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
bcrypt==4.0.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
ciso8601==2.2.0
click==8.1.3
dictdiffer==0.9.0
fastapi==0.89.1
fastapi-admin==1.0.4
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.3
orjson==3.8.3
pendulum==2.1.2
pycares==4.3.0
pycparser==2.21
pydantic==1.10.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
XlsxWriter==3.0.8
yarl==1.8.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1